% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgefinder.R
\name{clusteringCoef}
\alias{clusteringCoef}
\title{Calculate the clustering coefficient of each node.}
\usage{
clusteringCoef(A)
}
\arguments{
\item{A}{an adjacency Matrix (0/1).}
}
\value{
A vector with the clustering coefficient of each node.
}
\description{
Calculate the clustering coefficient of each node.
}
\examples{
\donttest{
   data(WT)
   WTres <- edgefinder(WT, ttl = "Wild Type")
   clusteringCoef(WTres$AdjMat)
}

}
