% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgefinder.R
\name{collapsedGraph}
\alias{collapsedGraph}
\title{Return an adjacency matrix after collapsing clusters into their central nodes.}
\usage{
collapsedGraph(A, clustersInfo)
}
\arguments{
\item{A}{An adjacency Matrix.}

\item{clustersInfo}{Obtained from graphComponents}
}
\value{
A weighted adjacency matrix between clusters and unclustered nodes.
}
\description{
Takes an object obtained from graphComponents and prints summary statistics.
}
\examples{
\donttest{
   data(WT)
   WTres <- edgefinder(WT, ttl = "Wild Type")
   rownames(WTres$AdjMat) = rownames(WT)
   WTComp <- graphComponents(WTres$AdjMat)
   Adj1 <- collapsedGraph(WTres$AdjMat, WTComp) > 0
   plotBitmapCC(Adj1,showMinDegree = 2)
}
}
