% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgefinder.R
\name{edgefinder}
\alias{edgefinder}
\title{Detect edges in co-expression datasets.}
\usage{
edgefinder(
  Exprs,
  BHthr = 0.01,
  rndseed = 112211,
  maxLen = 20000,
  LOvals = 30,
  ttl = "",
  trim = 0
)
}
\arguments{
\item{Exprs}{A numeric matrix with normalized gene expression data. Rows
correspond to genes, and columns correspond to samples.}

\item{BHthr}{the Benjamini-Hochberg fasle discovery rate threshold to be
used to determine which pairs are strongly correlated. Default=0.01.}

\item{rndseed}{The random seed used to select a subset of the pairs.}

\item{maxLen}{The maximum number of pairs that will be randomly selected
to fit the L2N model. Default=20000.}

\item{LOvals}{the maximum log-odds ratio to be used to determine the
cut-off points to declare which correlations are significant.
The program will check which log-odds ratio (1,2,...,LOvals) results in
FDR less than or equal to the user-specified BHthr. Default=30.}

\item{ttl}{Title for the fitted-model plot. Default=""}

\item{trim}{Fraction of extreme values to exclude from the fitted-model
plot. Default=0 (show all the data).}
}
\value{
A list with the following elements
\itemize{
\item{G} {The total number of genes.}
\item{p1} {The proportion of genes in the right mixture component (positively correlated.)}
\item{p2} {The proportion of genes in the left mixture component (negtively correlated.)}
\item{p0} {The proportion of genes in the null component (un-correlated.)}
\item{m0, m1, m2, s0, s1, s2} {The location and scale parameters of the three mixture components.}
\item {thrtable} {A table with 6 columns: posterior probability ratio (ppr) between the non-null components and the null component), the right component cutoff corresponding to the ppr, the left component cutoff, the estimated probability of Type-I errors, the estimated power, the estimated FDR.}
\item {LogOddsRatio} {The log-odds ratio that yields FDR closest to the desired level.}
\item {fitted} {The fitted model (as returned by the EM function).}
\item {rmse} {The root mean-squared error of the fitted model.}
\item {rt, lt} {The significant edges (from the right and left mixture components.)}
\item {AdjMat} {The (sparse) adjacency matrix with edges corresponding to rt, lt.}
}
}
\description{
Fit the L2N model to normalized correlation coefficients between pairs of
genes. The mixture model has three component - the null component follows
a normal distribution, and the two non-null components follow lognormal
distributions. An edge is in the graph if the correlation between the two
end-point genes is large enough and determined to be in one of the non-null
components.
}
\examples{
\donttest{
   data(WT)
   WTres <- edgefinder(WT, ttl = "Wild Type")
}
}
