% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgefinder.R
\name{plotCluster}
\alias{plotCluster}
\title{Plot cluster network}
\usage{
plotCluster(AdjMat, clustNo, clusterInfo = NULL)
}
\arguments{
\item{AdjMat}{An adjacency Matrix (0/1).}

\item{clustNo}{The chosen cluster.}

\item{clusterInfo}{Obtained from graphComponents.}
}
\description{
Plot a cluster network with all the nodes and edges - the central node is marked by a black circle. The radius of each point corresponds to its degree. The opacity corresponds to the percentage of edges from the node that is in the cluster (the darker it is, the larger the percentage of edges is within the cluster.) The distance from the center corresponds to the relative dissimilarity with the central node. This is computed as the number of neighbors the node and the central node do not have in common.
}
\examples{
\donttest{
   data(WT)
   WTres <- edgefinder(WT, ttl = "Wild Type")
   WTComp <- graphComponents(WTres$AdjMat)
   plotCluster(WTres$AdjMat, 5, WTComp)
}
}
