% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgefinder.R
\name{plotDegCC}
\alias{plotDegCC}
\title{Plot the degree of nodes versus the degree times the clustering coefficient.}
\usage{
plotDegCC(edgefinderobj, clusterInfo = NULL, highlightNodes = NULL)
}
\arguments{
\item{edgefinderobj}{The object (list) returned by edgefinder.}

\item{clusterInfo}{obtained from graphComponents. If not provided by the user, it will be computed on the fly.}

\item{highlightNodes}{A vector of node-numbers which will be shown in red. Default is NULL.}
}
\description{
The x-axis represents the number of neighbors of each node, and the y-axis represents the proportion of neighbors which are connected to each other.
}
\examples{
\donttest{
   data(WT)
   WTres <- edgefinder(WT, ttl = "Wild Type")
   WTComp <- graphComponents(WTres$AdjMat)
   plotDegCC(WTres,WTComp)
}
}
