% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgefinder.R
\name{plotMixture}
\alias{plotMixture}
\title{Plot the histogram of the data and the fitted mixture distribution.}
\usage{
plotMixture(fit.em, gof, ttl = "", trim = 0)
}
\arguments{
\item{fit.em}{The object (list) returned from the EM function with the parameter estimates for the L2N model.}

\item{gof}{The root mean-squared error of the fitted model (to appear in the title of the plot).}

\item{ttl}{The title of the plot (default="").}

\item{trim}{The proportion of extreme values on both sides of the distribution to eliminate from the plot (default=0.) This can be useful if a small number of values are so extreme, that the plot shows mostly the tails and a spike in the middle. Default=0.}
}
\description{
The function is called by the edgefinder function.
}
\examples{
\donttest{
   data(WT)
   WTres <- edgefinder(WT, ttl = "Wild Type")
   plotMixture(WTres$fitted, WTres$rmse)
}
}
