% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgefinder.R
\name{shortSummary}
\alias{shortSummary}
\title{Print a short summary of the fitted mixture model.}
\usage{
shortSummary(edgefinderobj)
}
\arguments{
\item{edgefinderobj}{The object (list) returned from the edgefinder function.}
}
\description{
Show the number of nodes, the number of possible and detected edges, the estimated proportion of positively/negatively correlated pairs, and the estimated false discovery rate.
}
\examples{
\donttest{
   data(WT)
   WTres <- edgefinder(WT, ttl = "Wild Type")
   shortSummary(WTres)
}
}
