% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgefinder.R
\name{summarizeClusters}
\alias{summarizeClusters}
\title{Show cluster characteristics.}
\usage{
summarizeClusters(clustersInfo)
}
\arguments{
\item{clustersInfo}{Obtained from graphComponents.}
}
\value{
A matrix with cluster number, number of nodes, and fivenum summaries for the degrees of nodes in the cluster, and the percentage of edges that are within the cluster.
}
\description{
Takes an object obtained from graphComponents and prints and returns summary statistics.
}
\examples{
\donttest{
   data(WT)
   WTres <- edgefinder(WT, ttl = "Wild Type")
   WTComp <- graphComponents(WTres$AdjMat)
   (summtab <- summarizeClusters(WTComp))
}
}
