% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgefinder.R
\docType{data}
\name{WT}
\alias{WT}
\title{Gene Expression data for the WildType group}
\format{
A matrix with 3454 rows and 15 columns
}
\usage{
data(WT)
}
\description{
WT is a matrix with normalized gene expression data containing 3454 differentially expressed genes (when compared with the duplication group) from 15 samples (columns) from the wild-type group.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/sites/GDSbrowser?acc=GDS4430}
}
\keyword{datasets}
